requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-gis
gunicorn
